MP3 ID3 V2.3 AENC, APIC, COMM, COMR tags

+++ src/main/java/com/aurawin/mp3/payload/EncryptedAudio.java
@@ -2,8 +2,8 @@ package com.aurawin.mp3.payload;
 
 public class EncryptedAudio {
      public String Owner;
-     public short PreviewStart;
-     public short PreviewLength;
+     public long PreviewStart;
+     public long PreviewLength;
      public byte[] Data;
 }
